requirements.txt
Flask==1.1.2
pdfkit==0.6.1
Werkzeug==1.0.1
mysql_connector_repackaged==0.3.1
passlib==1.7.2
python_dateutil==2.8.1
PyYAML==5.3.1
